% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chrome.R
\docType{data}
\name{Chrome}
\alias{Chrome}
\title{Chromium Concentration Data}
\format{{A data frame with 100 observations on the following 1 variables.
\describe{
\item{\code{conc}}{chromium concentration in individual workplaces mg/m^3}
}}}
\description{
Chromium air concentration in stainless steel welding workplaces.
}

