% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateVote.R
\docType{data}
\name{ClimateVote}
\alias{ClimateVote}
\title{Climate Vote Data}
\format{{ A data frame with 100 observations on the following 1 variables.
\describe{
\item{\code{ID}}{voter}
\item{\code{vote}}{1 = yes, 0 = no}
}}}
\description{
Votes on climate change referendum.
}

