% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAcceptability.R
\docType{data}
\name{EAcceptability}
\alias{EAcceptability}
\title{Environmental acceptability of hazards simulated opinion data}
\format{{ A data frame with 100 observations on the following 4 variables.
\describe{
\item{\code{score}}{acceptability of fraking}
\item{\code{gender}}{sex of respondent}
\item{\code{hincome}}{household income}
}}}
\description{
Simulated opinion data regarding the acceptability of fraking in the context of the environment used in the ordinal logit lab.
}


