% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chrome.R
\docType{data}
\name{FrackingYears}
\alias{FrackingYears}
\title{Fracking Acceptability Over Time}
\format{{A data frame with 120 observations on the following 2 variables.
\describe{
\item{\code{score}}{Fracking Acceptability Score Among US Citizens}
\item{\code{year}}{Survey Year}
}}}
\description{
How has perveived fracking acceptability changed over time among US citizens}

