% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IslandLizards.R
\docType{data}
\name{IslandsLizards}
\alias{IslandsLizards}
\title{Islands Lizards Data}
\format{{ A data frame with 19 observations on the following 2 variables.
\describe{
\item{\code{perimeterAreaRatio}}{Ratio of island perimeter to area}
\item{\code{presence}}{1 = present, 0 = absent}
}}}
\description{
Presence/absence of lizards on the California islands.
}

