% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\docType{data}
\name{Logistic}
\alias{Logistic}
\title{Logistic Data}
\format{{ A data frame with 50 observations on the following 3 variables.
\describe{
\item{\code{year}}{year}
\item{\code{PopulationSize}}{population size}
\item{\code{GrowthRate}}{population per capita growth rate}
}}}
\description{
Growth rates paired with population abundance.
}

