% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lynx.R
\docType{data}
\name{LynxFamilies}
\alias{LynxFamilies}
\title{Lynx Census and Harvest Data}
\format{{ A data frame with 19 observations on the following 3 variables.
\describe{
\item{\code{year}}{year survey was done}
\item{\code{census}}{number of lynx in RMU}
\item{\code{harvest}}{number of lynx harvested in RMU}
}}}
\description{
Abundance of Eurasian lynx family (\emph{Lynx lynx}) family groups over time for a single Reindeer Management Unit (RMU) in Sweden.
}
\references{
\url{http://www.swedishepa.se}
}

