% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N2OEmission.R
\docType{data}
\name{N2OEmission}
\alias{N2OEmission}
\title{N2O Emission Data}
\format{{ A data frame with 563 observations on the following 8 variables.
\describe{
\item{\code{fertilizer}}{type of fertilizer}
\item{\code{group}}{group}
\item{\code{carbon}}{carbon}
\item{\code{n.input}}{n.input}
\item{\code{emission}}{emission}
\item{\code{reps}}{reps}
\item{\code{group.index}}{group.index}
\item{\code{fert.index}}{fert.index}
}}}
\description{
N2O emissions from agricultural soils.
}

