% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chrome.R
\docType{data}
\name{PlantEnergy}
\alias{PlantEnergy}
\title{Plant Metabolizable Energy Data}
\format{{A data frame with 531 observations on the following 1 variables.
\describe{
\item{\code{conc}}{metabolizable energy (ME) in plants (k joule / g dry matter)}
}}}
\description{
Metabolizable energy concentration in woody and herbaceous plants.
}

