% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RichnessBirds.R
\docType{data}
\name{RichnessBirds}
\alias{RichnessBirds}
\title{Bird Species Richness Data}
\format{{ A data frame with 51 observations on the following 7 variables.
\describe{
\item{\code{rank}}{Ranking of species richness by state}
\item{\code{state}}{state}
\item{\code{st}}{state abbreviation}
\item{\code{species}}{number of bird species by state}
\item{\code{area}}{area of each state}
\item{\code{temp}}{mean tempterature in each state}
\item{\code{precip}}{mean precipiation in each state}
}}}
\description{
Species richness of bird species in the United States by state.
}

