\name{SESYNCBayes-package}
\alias{SESYNCBayes}
\docType{package}
\title{Bayesian Modeling for Socio-Environmental Data Package}
\description{Contains datasets and functions used in course labs and lectures.}
\details{
  \tabular{ll}{
    Package: \tab SESYNCBayes\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2016-02-04\cr
    License: \tab GPL-3\cr
  }
}
\author{
  Tom Hobbs, Mary Collins, and Christian Che-Castaldo

  Maintainer: Christian Che-Castaldo <checastaldo@icloud.com>}
\references{NA}
\keyword{ package }
\seealso{}
\examples{}
