% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteCarbon.R
\docType{data}
\name{SiteCarbon}
\alias{SiteCarbon}
\title{Site Carbon Data}
\format{{ A data frame with 107 observations on the following 4 variables.
\describe{
\item{\code{group}}{group}
\item{\code{mean}}{mean}
\item{\code{sd}}{sd}
\item{\code{group.index}}{group.index}
}}}
\description{
Site carbon data
}

