% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolsticeTemp.R
\docType{data}
\name{SolsticeTemp}
\alias{SolsticeTemp}
\title{Solstice Temperature Data}
\format{{ A data frame with 50 observations on the following 2 variables.
\describe{
\item{\code{ID}}{year}
\item{\code{temp}}{temperature (F)}
}}}
\description{
Solstice temperature data in Maryland.
}

