% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissBirds.R
\docType{data}
\name{SwissBirds}
\alias{SwissBirds}
\title{Swiss Birds Occupancy Data}
\format{{ A data frame with 237 observations on the following 5 variables.
\describe{
\item{\code{quadrat}}{1 km sq quadrat number}
\item{\code{numberVisits}}{number of surveys}
\item{\code{numberDetections}}{number of times birds were detected}
\item{\code{elevation}}{site elevation (m)}
\item{\code{forestCanopy}}{percent forest cover at each site}
}}}
\description{
Presence/absence of the willow tit (\emph{Parus montanus}) in the Swiss Alps taken from the Monitoring Häufige Brutvögel (MHB) common breeding bird survey.
}
\references{
\url{http://www.vogelwarte.ch/en/projects/monitoring/monitoring-common-breeding-birds.html}
}

